requirements.txt
python-dotenv
python-multipart
fastapi
uvicorn
sqlmodel
stripe
requests